+++ Tests/intent/Application.Identity.AccountController.TestApplication/Application.Identity.AccountController.Infrastructure/Persistence/ApplicationDbContext.cs
@@ -1,4 +1,6 @@
 using Application.Identity.AccountController.Domain.Common.Interfaces;
+using Application.Identity.AccountController.Domain.Entities;
+using Application.Identity.AccountController.Infrastructure.Persistence.Configurations;
 using Intent.RoslynWeaver.Attributes;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
@@ -9,17 +11,20 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Application.Identity.AccountController.Infrastructure.Persistence
 {
-    public class ApplicationDbContext : IdentityDbContext<IdentityUser>, IUnitOfWork
+    public class ApplicationDbContext : IdentityDbContext<ApplicationIdentityUser>, IUnitOfWork
     {
         public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options)
         {
         }
+
+        public DbSet<ApplicationIdentityUser> ApplicationIdentityUsers { get; set; }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
             base.OnModelCreating(modelBuilder);
 
             ConfigureModel(modelBuilder);
+            modelBuilder.ApplyConfiguration(new ApplicationIdentityUserConfiguration());
         }
 
         [IntentManaged(Mode.Ignore)]