+++ Modules/Intent.Modules.AspNetCore.Identity.AccountController/release-notes.md
@@ -1,6 +1,15 @@
+﻿### Version 2.0.0
+
+- Added Refresh Token functionality.
+- `AccountController` class refactored to have `TokenService` separately.
+
 ### Version 1.1.2
 
 - Updated dependencies and supported client versions to prevent warnings when used with Intent Architect 4.x.
+
+### Version 1.1.1
+
+- Update: Role Claims are now introduced.
 
 ### Version 1.1.0